kwin: fix aurorae (QML themes)
The plugin for QML themes wasn't being loaded because kwin hacks the
search path and upstream forgot to use FollowSymlinks with
QDirIterator. That seems to be a common problem; I'm going to search the
code for all uses of QDirIterator next. This fixes the shiny new Breeze
KWin theme, but it's still sloooooow, so don't use it.