qgroundcontrol: broken with Qt 5.8
This package has been broken for some time because it relies on private headers
from Qt 5.5, which breaks Qt's API and ABI guarantees. Upstream is aware of the
problem and does not intend to fix it:
https://github.com/mavlink/qgroundcontrol
This suggests that the package is NOT suitable for inclusion in Nixpkgs. Users
who want this package should download upstream's binary release and run it in a
chroot.

The Nixpkgs expression also seems to have been abandoned by its maintainer. It
has never been updated since it was originally added, despite 11 new releases in
that time.