qt-5: 5.15.0 -> 5.15.2
This is a maintenance release of the Qt 5.15 release series. It focuses on
bug and security fixes. The feature that made me update this package was
the broken u2f support in QtWebEngine when used together with
Qutebrowser. It previously (on older Qt versions) used to work but
stopped working with the switch to 5.15.

As the changelog is rather extensive the reader is advised to inspect
[0] to checkout the changes to each of the components of their
interest.

This also fixes a security issue with QtWebEngine based browsers that
could be crashed by visiting a website that invokes certain WebRTC
features [1].

[0] https://wiki.qt.io/Qt_5.15.2_Change_Files
[1] https://bugreports.qt.io/browse/QTBUG-86752

(cherry picked from commit dd6446e17f9ff7498588c39c88a0bd1bd8110d9e)